+++ lib/view_model/dashboard_view_model/dashboard_view_model.dart
@@ -32,7 +32,7 @@ class DashboardViewModel extends GetxController {
     super.onInit();
     currentTab = 0;
     currentScreen = const HomePage();
-    Get.put(HomeController());
+    Get.put(HomeViewModel());
     update();;
   }
