+++ lib/view_model/home_view_model/home_binding.dart
@@ -4,6 +4,6 @@ import 'package:smart_banking/view_model/home_view_model/home_view_model.dart';
 class HomedBinding extends Bindings {
   @override
   void dependencies() {
-    Get.put(() => HomeController(), permanent: false);
+    Get.put(() => HomeViewModel(), permanent: false);
   }
 }
